remove gitlab2zenodo
The CI hook to update Zenodo when a new version was created didn't work
because of a file too large error. There was no diagnostic information
about which file was too large, and it looked like all the files were
within the limit required. Going to try pushing to GitHub to see if the
GitHub/Zenodo integration works better.